Using longs instead of ints for wordcount.

+++ trunk/samplemodels/wordcount/src/main/java/es/tid/bdp/samples/wordcount/WordCountReducer.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mapreduce.Reducer;
 
 /**
@@ -12,14 +13,14 @@ import org.apache.hadoop.mapreduce.Reducer;
  * @author logc
  */
 public class WordCountReducer
-        extends Reducer<Text, IntWritable, Text, IntWritable> {
+        extends Reducer<Text, IntWritable, Text, LongWritable> {
     @Override
     public void reduce(Text key, Iterable<IntWritable> values, Context context) 
           throws IOException, InterruptedException {
-        int sum = 0;
+        long sum = 0L;
         for (IntWritable val : values) {
             sum += val.get();
         }
-        context.write(key, new IntWritable(sum));
+        context.write(key, new LongWritable(sum));
     }
 }
